server/requirements.txt
crewai
yfinance
requests
python-dotenv
langchain-google-genai
